% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Golay_Raster.R
\name{Golay_Raster}
\alias{Golay_Raster}
\title{Gapfill Raster using Savitzky-Golay filtering}
\usage{
Golay_Raster(path, path_result)
}
\arguments{
\item{path}{The path contains the images (conditions over 30 Images)}

\item{path_resule}{The path contains the result images}
}
\value{
Result image
}
\description{
Use Savitzky-Golay time filtering algorithm to fill
satellite images lost image elements.
}
