% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQR_Raster.R
\name{IQR_Raster}
\alias{IQR_Raster}
\title{IQR Raster}
\usage{
IQR_Raster(x, type_img)
}
\arguments{
\item{x}{file ảnh cần lọc giá trị ngoại lệ}
}
\value{
kết quả ảnh
}
\description{
Lọc giá trị ngoại lệ dựa trên công thức IQR (InterQuartile Range)
}
