% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mask_Multi_Raster.R
\name{Mask_Multi_Raster}
\alias{Mask_Multi_Raster}
\title{Mask Multifile Raster}
\usage{
Mask_Multi_Raster(path, path_result, vector_shp)
}
\arguments{
\item{path}{đường dẫn chứa ảnh}

\item{path_result}{đường dẫn chứa kết quả ảnh}

\item{vector_shp}{dữ liệu vector (ESRI Shapefiles)}
}
\value{
kết quả ảnh
}
\description{
Tạo mask và cắt nhiều ảnh Raster
}
