% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mean_Raster.R
\name{Mean_Raster}
\alias{Mean_Raster}
\title{Calculate Mean Raster}
\usage{
Mean_Raster(path)
}
\arguments{
\item{path}{The path contains the Raster}
}
\value{
Result Images
}
\description{
Calculate Raster of TVDI index
}
