% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TVDI_Largefiles_process.R
\name{TVDI_Largefiles_process}
\alias{TVDI_Largefiles_process}
\title{Calculate TVDI and Export TVDI Images}
\usage{
TVDI_Largefiles_process(path_NDVI, path_LST, path)
}
\arguments{
\item{path_NDVI}{Path contains NDVI images}

\item{path_LST}{Path contains LST images}

\item{path}{Path contains TVDI directory}
}
\value{
Results TVDI
}
\description{
Set the function of determining TVDI through LST and NDVI on all image points.
Then export the result as Raster TVDI.
Note: Firstly, this function will process large files If you have small or medium files, you can
use 'TVDI_process' in order to process small and medium files. Secondly, I have divided roughly in
this calculation has a distance of 0.05, can be adjusted by downloading this function and changing
to suit the data
}
