% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Golay_Raster.R
\name{Golay_Raster}
\alias{Golay_Raster}
\title{Gapfill Raster using Savitzky-Golay filtering}
\usage{
Golay_Raster(path, path_result)
}
\arguments{
\item{path}{đường dẫn chứa ảnh (điều kiện trên 30 thời điểm ảnh)}

\item{path_resule}{đường dẫn chứa kết quả ảnh}
}
\value{
kết quả ảnh
}
\description{
Sử dụng thuật toán lọc thời gian Savitzky-Golay để lắp đầy
ảnh vệ tinh bị mất các phần tử ảnh.
}
