% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQR_Raster.R
\name{IQR_Raster}
\alias{IQR_Raster}
\title{IQR Raster}
\usage{
IQR_Raster(x, type_img)
}
\arguments{
\item{x}{Image file to filter out exception value}

\item{type_img}{Type of Image}
}
\value{
Result image
}
\description{
Filter outliers based on formula IQR (InterQuartile Range)
}
