% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mask_Multi_Raster.R
\name{Mask_Multi_Raster}
\alias{Mask_Multi_Raster}
\title{Mask Multifile Raster}
\usage{
Mask_Multi_Raster(path, path_result, vector_shp)
}
\arguments{
\item{path}{The path contains the images}

\item{path_result}{The path contains the result images}

\item{vector_shp}{vector data (ESRI Shapefiles)}
}
\value{
Result images
}
\description{
Create a mask and crop multiple raster images
}
