% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TVDI_process.R
\name{TVDI_process}
\alias{TVDI_process}
\title{Calculate TVDI and Export TVDI Images}
\usage{
TVDI_process(path_NDVI, path_LST, path)
}
\arguments{
\item{list_NDVI}{List of NDVI images}

\item{list_LST}{List of LST images}
}
\value{
Results TVDI
}
\description{
Set the function of determining TVDI through LST and NDVI on all image points.
Then export the result as Raster TVDI.
}
