% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TVDI_process.R
\name{TVDI_process}
\alias{TVDI_process}
\title{Calculate TVDI and Export TVDI Images}
\usage{
TVDI_process(path_NDVI, path_LST, path)
}
\arguments{
\item{list_NDVI}{Danh sách các ảnh NDVI}

\item{list_LST}{Danh sách các ảnh LST}
}
\value{
kết quả ảnh TVDI
}
\description{
Lập hàm xác định TVDI thông qua 2 chỉ số LST và NDVI trên toàn bộ các thời điểm ảnh.
Sau đó xuất kết quả dưới dạng Raster TVDI.
}
